+++ R/001.bquote.R
@@ -1,5 +1,0 @@
-#' @importFrom utils globalVariables
-globalVariables(c(".", ".."))
-
-bquote_compile <- import_future("bquote_compile")
-bquote_apply <- import_future("bquote_apply")