+++ R/zzz.R
@@ -1,10 +1,4 @@
-## To be cached by .onLoad()
-FutureRegistry <- NULL
-assertOwner <- NULL
-
 .onLoad <- function(libname, pkgname) {
-  ## Import private functions from 'future'
-  FutureRegistry <<- import_future("FutureRegistry")
-  assertOwner <<- import_future("assertOwner")
+  import_future_functions()
 }